Updated BPR4 Project

+++ FH/Python/BPR4/MHCII/main.py
@@ -18,36 +18,37 @@ PDBDIR = 'PDB//'
 LOWER_BOUND = 9
 UPPER_BOUND = 25
 
-
-
-
 # Extracts relevant sequences and their IC50 statistics from the specified CSV.
 def get_assay_entries(assay_filename):
     entries = dict()
     assay_filename_noext = '.'.join(assay_filename.split('.')[:-1])
     original_ext = assay_filename.split('.')[-1]
+    cache_name = assay_filename_noext +"_"+str(LOWER_BOUND)+'-'+str(UPPER_BOUND) + '.cache'
 
     if not os.path.exists(assay_filename):
         print(assay_filename + ": specified file could not be found.")
         return []
 
+    # if assay file is a zip, unzip it first
     if original_ext=='zip':
         with zipfile.ZipFile(assay_filename, 'r') as zip_ref:
             zip_ref.extractall('.')
             assay_filename = assay_filename_noext+'.csv'
 
-    if not os.path.exists(assay_filename_noext + '.cache'):
+    if not os.path.exists(cache_name):
+        num_lines = 0
+
         with open(assay_filename, "r", encoding='utf-8') as iedb_file:
             # skip first two lines
             line = iedb_file.readline()
             line = iedb_file.readline()
+            linecount = iedb_file
 
             csvreader = csv.reader(iedb_file, delimiter=',', quotechar='"')
 
             for row in csvreader:
+                num_lines += 1
                 seq = row[11]
-                # seq = re.sub('\u02d9', '', seq)
-                # seq = re.sub(' \+.*$', '', seq)
 
                 if len(seq) >= LOWER_BOUND and len(seq) <= UPPER_BOUND:
                     tmp = list()
@@ -62,23 +63,25 @@ def get_assay_entries(assay_filename):
                     entries[seq].append(tmp)
 
         # after processing of the file, cache the relevant entries
-        with open(assay_filename_noext + '.cache', 'w') as seq_file:
+        with open(cache_name, 'w') as seq_file:
             for entry in entries.keys():
                 for subentry in entries[entry]:
                     seq_file.write(entry + ',' + ','.join(subentry) + '\n')
     else:
-        with open(assay_filename_noext + '.cache') as seq_file:
+        # if cache file exists, read the data from it instead
+        with open(cache_name) as seq_file:
             for line in seq_file.readlines():
                 data = line.split(',')
                 if data[0] not in entries.keys():
                     entries[data[0]] = list()
-            tmp = list()
-            tmp.append(data[1])
-            tmp.append(data[2])
-            tmp.append(data[3])
-            tmp.append(data[4])
-            entries[data[0]].append(tmp)
+                tmp = list()
+                tmp.append(data[1])
+                tmp.append(data[2])
+                tmp.append(data[3])
+                tmp.append(data[4])
+                entries[data[0]].append(tmp)
 
+    # if assays were zipped, delete the extracted content to save data
     if original_ext == 'zip':
         if os.path.exists(assay_filename):
             os.remove(assay_filename)
@@ -103,13 +106,23 @@ def map_to_pdb(entries, seqres_filename):
             line = line.strip()
             matches = re.match('^>([^_]+).*mol:protein.*length:(\d+)\s+(.*)$', line)
             if skipper == False and line[0] != '>':
-                if line in entries.keys():
+
+                found = False
+                found_seq = ''
+                for entry_seq in entries.keys():
+                    if entry_seq in line:
+                        found = True
+                        found_seq = entry_seq
+                        break
+
+                if found:
                     pdb_ids.add(acc_number)
                     if acc_number not in mapped_infos.keys():
                         mapped_infos[acc_number] = list()
                     mapped_infos[acc_number].append(line)
-                    mapped_infos[acc_number].append(entries[line])
+                    mapped_infos[acc_number].append(entries[found_seq])
 
+            # if line is header and length not within bounds, skip the next line
             elif line[0] == '>' and matches:
                 length = int(matches.group(2))
                 acc_number = matches.group(1)
@@ -132,7 +145,7 @@ def map_to_pdb(entries, seqres_filename):
 def download_pdb(pdb_id):
     # If file already exists, skip the download.
     base_url = 'http://files.rcsb.org/download/'
-    if not os.path.exists(PDBDIR + pdb_id + '.pdb'):
+    if not os.path.exists(PDBDIR + pdb_id + '.pdb') and not os.path.exists(PDBDIR + pdb_id + '.cif'):
         print("Downloading "+base_url+pdb_id)
         try:  # try .pdb
             urllib.request.urlretrieve(base_url + pdb_id + '.pdb', PDBDIR + pdb_id + '.pdb')
@@ -179,5 +192,13 @@ def download_pdbs(pdb_ids):
 
 if __name__ == '__main__':
     iedb_entries = get_assay_entries('mhc_ligand_table_export_1615816328.zip')
+    print("Finished extracting IEDB data.")
+
+    print("Now mapping IEDB to PDB. This could take a while...")
     pdb_ids = map_to_pdb(iedb_entries, 'pdb_seqres.txt.gz')
-    download_pdbs(pdb_ids)
+    print(str(len(pdb_ids)) + " PDB IDs were successfully mapped to " + str(len(iedb_entries)) + " IEDB assay entries.")
+    print("Inspect mapped_info.txt for more details.\n")
+
+    ans = input("Do you also want to download all mapped PDB files? (y/n)")
+    if ans.lower() in ['y','yes']:
+        download_pdbs(pdb_ids)
